requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0